+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.30
 * New function `local_growth_rates`
+* New function `match_attractors!`
 
 # 1.29
 * Performance improvements for basins of attraction